JPA GenreDao and GenreDaoTests

+++ library/src/main/java/ru/otus/spring01/library/dao/GenreDao.java
@@ -1,13 +1,12 @@
 package ru.otus.spring01.library.dao;
 
 import ru.otus.spring01.library.domain.Genre;
-import ru.otus.spring01.library.domain.Person;
 
 import java.util.List;
 import java.util.UUID;
 
 public interface GenreDao {
-    int count();
+    Long count();
 
     void insert(Genre genre);
